Avoid integer overflow on long-running calculations

+++ src/elfcode.h
@@ -1,3 +1,4 @@
 #include <stdbool.h>
-int run(int *r, const int ip, const int *program, const int len, int max,
-        bool print, int *trace);
+#include <stdint.h>
+int64_t run(int *r, const int ip, const int *program, const int len, int max,
+            bool print, int *trace);
